+++ app/lib/ui/components/link/link_customization.dart
@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_widget_state.dart';
 import 'package:ouds_flutter_demo/ui/components/link/link_enum.dart';
+import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_widget_state.dart';
 
 /// Section for InheritedWidget to pass data down the widget tree
 class _LinkCustomization extends InheritedWidget {
-  const _LinkCustomization({
-    required super.child,
-    required this.data,
-  });
+  const _LinkCustomization({required super.child, required this.data});
 
   final LinkCustomizationState data;
 
@@ -17,10 +14,7 @@ class _LinkCustomization extends InheritedWidget {
 
 /// Main Widget class for tag customization
 class LinkCustomization extends StatefulWidget {
-  const LinkCustomization({
-    super.key,
-    required this.child,
-  });
+  const LinkCustomization({super.key, required this.child});
 
   final Widget child;
 
@@ -28,15 +22,19 @@ class LinkCustomization extends StatefulWidget {
   LinkCustomizationState createState() => LinkCustomizationState();
 
   static LinkCustomizationState? of(BuildContext context) {
-    return (context.dependOnInheritedWidgetOfExactType<_LinkCustomization>())?.data;
+    return (context.dependOnInheritedWidgetOfExactType<_LinkCustomization>())
+        ?.data;
   }
 }
 
 /// tag customization state management
-class LinkCustomizationState extends CustomizationWidgetState<LinkCustomization> {
+class LinkCustomizationState
+    extends CustomizationWidgetState<LinkCustomization> {
   late final LayoutState layoutState;
   late final LabelTextState labelTextState;
   late final SizeState sizeState;
+  late final DensityState densityState;
+  late final TintedState tintedState;
 
   @override
   void initState() {
@@ -44,6 +42,8 @@ class LinkCustomizationState extends CustomizationWidgetState<LinkCustomization>
     layoutState = LayoutState(setState);
     labelTextState = LabelTextState(setState);
     sizeState = SizeState(setState);
+    densityState = DensityState(setState);
+    tintedState = TintedState(setState);
   }
 
   LinkEnumLayout get selectedLayout => layoutState.selected;
@@ -55,16 +55,17 @@ class LinkCustomizationState extends CustomizationWidgetState<LinkCustomization>
   LinkEnumSize get selectedSize => sizeState.selected;
   set selectedSize(LinkEnumSize value) => sizeState.selected = value;
 
+  LinkEnumDensity get selectedDensity => densityState.selected;
+  set selectedDensity(LinkEnumDensity value) => densityState.selected = value;
+
+  bool get isTinted => tintedState.value;
+  set isTinted(bool value) => tintedState.value = value;
 
   @override
   Widget build(BuildContext context) {
-    return _LinkCustomization(
-      data: this,
-      child: widget.child,
-    );
+    return _LinkCustomization(data: this, child: widget.child);
   }
 }
-
 
 /// Layout State Management
 class LayoutState {
@@ -74,8 +75,9 @@ class LayoutState {
   final List<LinkEnumLayout> _layout = [
     LinkEnumLayout.textOnly,
     LinkEnumLayout.textAndIcon,
-    LinkEnumLayout.back,
+    LinkEnumLayout.previous,
     LinkEnumLayout.next,
+    LinkEnumLayout.external,
   ];
 
   List<LinkEnumLayout> get list => _layout;
@@ -110,10 +112,7 @@ class SizeState {
 
   final void Function(void Function()) _setState;
 
-  List<LinkEnumSize> _sizeList = [
-    LinkEnumSize.defaultSize,
-    LinkEnumSize.small
-  ];
+  List<LinkEnumSize> _sizeList = [LinkEnumSize.defaultSize, LinkEnumSize.small];
   LinkEnumSize _selectedSize = LinkEnumSize.defaultSize;
 
   List<LinkEnumSize> get list => _sizeList;
@@ -127,6 +126,51 @@ class SizeState {
   set selected(LinkEnumSize newValue) {
     _setState(() {
       _selectedSize = newValue;
+    });
+  }
+}
+
+/// Tinted State Management
+///
+/// Controls whether the icon displayed by [OudsLink.icon] should be tinted
+/// with the theme color, or shown with its original colors.
+class TintedState {
+  TintedState(this._setState);
+
+  final void Function(void Function()) _setState;
+  bool _tinted = true;
+
+  bool get value => _tinted;
+  set value(bool newValue) {
+    _setState(() {
+      _tinted = newValue;
+    });
+  }
+}
+
+/// Density State Management
+class DensityState {
+  DensityState(this._setState);
+
+  final void Function(void Function()) _setState;
+
+  List<LinkEnumDensity> _densityList = [
+    LinkEnumDensity.defaultDensity,
+    LinkEnumDensity.compact,
+  ];
+  LinkEnumDensity _selectedDensity = LinkEnumDensity.defaultDensity;
+
+  List<LinkEnumDensity> get list => _densityList;
+  set list(List<LinkEnumDensity> newList) {
+    _setState(() {
+      _densityList = newList;
+    });
+  }
+
+  LinkEnumDensity get selected => _selectedDensity;
+  set selected(LinkEnumDensity newValue) {
+    _setState(() {
+      _selectedDensity = newValue;
     });
   }
 }