+++ ouds_core/lib/components/form_input/password_input/ouds_password_input.dart
@@ -98,7 +98,7 @@ class OudsPasswordInput extends StatefulWidget {
     final theme = OudsTheme.of(context);
     return SvgPicture.asset(
       excludeFromSemantics: true,
-      AppAssets.icons.communicationSecurityAndSafetyLock,
+      AppAssets.icons.communicationSecurityAndSafetyLockClosed,
       package: OudsTheme.of(context).packageName,
       fit: BoxFit.contain,
       height: theme.componentsTokens(context).textInput.sizeLeadingIcon,
@@ -602,7 +602,7 @@ class _OudsPasswordInputState extends State<OudsPasswordInput> {
               appearance: OudsButtonAppearance.minimal,
               icon: _isPasswordHidden
                   ? AppAssets.icons.accessibilityAccessibilityVision
-                  : AppAssets.icons.settingsAndToolsHide,
+                  : AppAssets.icons.settingsAndToolsAccessibilityHide,
               onPressed: (widget.enabled ?? true && !(widget.readOnly ?? false))
                   ? _toggleIcon
                   : null,