+++ ouds_core/lib/components/link/internal/ouds_link_size_modifier.dart
@@ -26,13 +26,24 @@ class OudsLinkSizeModifier {
 
   OudsLinkSizeModifier(this.context);
 
-  /// Retrieves the min width and min height (double) for the Link based on the provided size enum.
-  Map<String, double?> getMinWidthAndHeight(OudsLinkSize? size) {
+  /// Retrieves the min width and min height (double) for the Link based on the provided size and density.
+  Map<String, double?> getMinWidthAndHeight(
+    OudsLinkSize size,
+    OudsLinkDensity density,
+  ) {
     final theme = OudsTheme.of(context).componentsTokens(context).link;
 
+    if (density == OudsLinkDensity.compact) {
+      return {
+        OudsLinkDimensions.width.name: theme.sizeMinWidth,
+        OudsLinkDimensions.height.name: theme.sizeMinHeightCompactDensity,
+      };
+    }
+
+    //default density
     if (size == OudsLinkSize.small) {
       return {
-        OudsLinkDimensions.width.name: theme.sizeMinWidthSmall,
+        OudsLinkDimensions.width.name: theme.sizeMinWidth,
         OudsLinkDimensions.height.name: theme.sizeMinHeightSmall,
       };
     } else {
@@ -44,7 +55,7 @@ class OudsLinkSizeModifier {
   }
 
   /// Retrieves the width and height of icon/ chevron (double) for the Link based on the provided size enum.
-  Map<String, double?> getIconSize(OudsLinkSize? size) {
+  Map<String, double?> getIconSize(OudsLinkSize size) {
     final theme = OudsTheme.of(context).componentsTokens(context).link;
 
     if (size == OudsLinkSize.small) {
@@ -61,9 +72,25 @@ class OudsLinkSizeModifier {
   }
 
   /// Retrieves the spaceColumnGap (double) for the Link based on the provided size enum.
-  double? getSizeColumnGap(OudsLinkSize? size, OudsLinkLayout layout) {
+  // the param layout will be removed
+  double? getSizeColumnGap(
+    OudsLinkSize size,
+    OudsLinkLayout layout,
+    OudsLinkIndicator? indicator,
+    String? icon,
+  ) {
     final linkToken = OudsTheme.of(context).componentsTokens(context).link;
 
+    if (indicator == null && icon != null) {
+      //textAndIcon
+      return size == OudsLinkSize.small
+          ? linkToken.spaceColumnGapIconSmall
+          : linkToken.spaceColumnGapIconDefault;
+    } else if (indicator != null) {
+      return size == OudsLinkSize.small
+          ? linkToken.spaceColumnGapChevronSmall
+          : linkToken.spaceColumnGapChevronDefault;
+    }
     switch (layout) {
       case OudsLinkLayout.textAndIcon:
         return size == OudsLinkSize.small
@@ -78,6 +105,33 @@ class OudsLinkSizeModifier {
         return size == OudsLinkSize.small
             ? linkToken.spaceColumnGapChevronSmall
             : linkToken.spaceColumnGapChevronDefault;
+    }
+  }
+
+  /// Retrieves the block/inline padding ([EdgeInsetsDirectional]) for the Link based on the provided size and density.
+  EdgeInsetsDirectional getPadding(OudsLinkSize size, OudsLinkDensity density) {
+    final linkToken = OudsTheme.of(context).componentsTokens(context).link;
+
+    if (density == OudsLinkDensity.compact) {
+      return EdgeInsetsDirectional.symmetric(
+        vertical: size == OudsLinkSize.small
+            ? linkToken.spacePaddingBlockCompactDensitySmall
+            : linkToken.spacePaddingBlockCompactDensityDefault,
+        horizontal: linkToken.spacePaddingInline,
+      );
+    }
+
+    // density = Default
+    if (size == OudsLinkSize.small) {
+      return EdgeInsetsDirectional.symmetric(
+        vertical: linkToken.spacePaddingBlockSmall,
+        horizontal: linkToken.spacePaddingInline,
+      );
+    } else {
+      return EdgeInsetsDirectional.symmetric(
+        vertical: linkToken.spacePaddingBlockDefault,
+        horizontal: linkToken.spacePaddingInline,
+      );
     }
   }
 }